int += float这种代码也能写出来的啊 你妈死了

+++ Game2/timer.h
@@ -10,7 +10,7 @@ public:
 		shotted = false;
 	}
 
-	void set_wait_time(int val) {
+	void set_wait_time(float val) {
 		wait_time = val;
 	}
 
@@ -44,8 +44,8 @@ public:
 		}
 	}
 private:
-	int pass_time = 0;						//已过时间
-	int wait_time = 0;						// 等待时间
+	float pass_time = 0;					//已过时间
+	float wait_time = 0;					// 等待时间
 	bool paused = false;					//是否暂停
 	bool shotted = false;					//是否触发
 	bool one_shoot = false;					//单次触发
